requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
pdfplumber
PyMuPDF
Pillow
pysqlite3-binary